Update zm-alarm_02.py
- Added the option to analyze more frames (default e)
- Added the option to independently choose which camera ID will send emails
- Added a heart beat timestamp file that can be monitored and restarted (crontab) should the script crash for any reason

+++ zm-alarm_02.py
@@ -16,6 +16,7 @@ import os
 import shutil
 import glob
 import imghdr
+import re
 
 # imports for opencv
 import numpy as np
@@ -44,7 +45,13 @@ import cv2
 # Camera ID's to perform object detection
 #
 cameras = "3,12,13,9,8"
-
+#
+# Camera ID's to send emails
+emails = "9"
+#
+# How many frames to analyse for each alarm
+frame_analysis = 4 # default 3. Increase for better accuracy but analysis takes longer
+#
 # ZM authentification
 zm_user="admin"
 zm_pass="admin"
@@ -55,6 +62,7 @@ db_cred = {"host":"localhost", "user":"zmuser", "password":"zmpass", "database":
 # Paths
 path_pic = "/home/user/zm_ai/pic/" # tmp folder for img to be analyzed
 path_opencv = "/home/user/zm_ai/opencv4/" # locations of files coco.names, yolov4.cfg, yolov4.weights
+path_hb = "/tmp/" # heartbeat path for timestamp file. Use crontab and external shell script to scan and restart if timestamp is not current. 
 #
 # ssmtp has to be setup and working under normal user
 # enter email to send notification to
@@ -81,6 +89,7 @@ delay = 1
 # log file to keep record.
 log_file = "/var/www/html/zm-alarm_output.log"
 log_enable = 1 # enable logging to file (1) or not (0)
+hb_enable = 1 # enable heartbeat
 
 #################################################
 # End of settings section
@@ -95,7 +104,6 @@ log_enable = 1 # enable logging to file (1) or not (0)
 monitors = {}
 alarms = []
 cookies = ""
-script_start = time.time()
 
 for index in cameras.split(","):
 	monitors[index] = {'State':-1,'St':-1,'Et':-1, 'Set': 0}
@@ -111,15 +119,16 @@ def main():
 # initialize empty list to store alarm events 
 	global alarms, cookies, script_start, mydb
 	retry = 0
-	ai_status = "Off"
+# init misc local variables
+	current_time = time.time()
+	current_time_hb = time.time()
+	heartbeat() # initiate hb file
 #init database
 	mydb = init_mysql()
 
-# clear screen and reset log file on start
+# clear screen
 	os.system('clear')
-	if os.path.exists(log_file):
-		os.remove(log_file)
-		
+	
 # Initialize folders
 	if not os.path.isdir(path_pic):
 		os.makedirs(path_pic)
@@ -130,15 +139,12 @@ def main():
 	if not os.path.isdir(path_opencv):
 		printLog(tstamp(), " Verify path and files coco.names, yolov4.cfg, yolov4.weights")
 		exit()
+		
+	if not os.path.isdir(path_hb) and hb_enable == 1:
+		printLog(tstamp(), " Verify heartbeat path")
+		exit()
 
-	printLog(tstamp(), "zm-alarm monitoring started")
-#
-#	Verify if opencv4 function is running in multiprocessing
-	if(ai_opencv4.is_alive()):
-		ai_status = "Running"
-	else:
-		ai_status = "Off - problem somewhere"
-	printLog(tstamp(), ai_status)
+	printLog(tstamp(), "zm-alarm monitoring started:", os.path.basename(__file__))
 #
 # Call login function to get token and store cookie. Login to be reinitialized every hour
 #
@@ -176,11 +182,17 @@ def main():
 		if alarms:
 			retry = sql_query(retry)
 		# Re initialize cookies and token every hour
-		tm = time.time() - script_start
+		tm = time.time() - current_time
 		if(tm > 3600):
 			a_token, cookies = login()
-			script_start = time.time()
+			current_time = time.time()
 		# printLog(tstamp(), "Reset cookies")
+		# Touch heartbeat file every minute
+		tm_hb = time.time() - current_time_hb
+		if(tm_hb > 60):
+			current_time_hb = time.time()
+			heartbeat()
+
 		time.sleep(delay)
 		
 #################################################
@@ -200,15 +212,22 @@ def login():
 # get state of monitors with ZMU. 3 = alarm, 1 or 5 is idle
 def monitor_state(a_token):
 	for index in monitors:
-		cmd = ['zmu', '-m', index , '-T', a_token, '-e', '-s',]
-		result = subprocess.run(cmd, stdout=subprocess.PIPE,stderr=subprocess.DEVNULL)
-		x=result.stdout.decode('UTF-8').strip().split(" ")
-		# catch error. Sometimes (rarely) zmu returns something else
-		if x[0].isdigit():
+		cmd = ['zmu', '-m', index, '-T', a_token, '-e', '-s']
+		try:
+			result = subprocess.run(cmd, stdout=subprocess.PIPE, stderr=subprocess.DEVNULL, check=True)
+			x = result.stdout.decode('UTF-8').strip().split(" ")
 			monitors[index]['State'] = int(x[0])
-		else:
-			# default to idle
-			monitors[index]['State'] = 1
+		except subprocess.CalledProcessError as e:
+			# Handle subprocess error
+			print(f"Error running subprocess for monitor {index}. Return code: {e.returncode}")
+			monitors[index]['State'] = 1  # Default to idle
+		except (IndexError, ValueError) as e:
+			# Handle index or value errors
+			print(f"Error parsing response for monitor {index}. {e}")
+			monitors[index]['State'] = 1  # Default to idle
+
+
+
 # timestamp
 def tstamp():
 	x = datetime.datetime.now()
@@ -218,15 +237,17 @@ def tstamp():
 # includes retry code when SQL query doesn't behave
 def sql_query(retry):
 
-	global mydb
+	global mydb, frame_analysis
 	if mydb.open:
 		m_id = alarms[0][0]
 		st = alarms[0][1]
 		et =alarms[0][2]
+		# timstamp may be a bit late compared to zoneminder alarm so move the start time by 1s default)
+		st = (datetime.datetime.strptime(st, "%Y-%m-%d %H:%M:%S") + datetime.timedelta(seconds=-1)).strftime("%Y-%m-%d %H:%M:%S")
 		# if alarms in queue
 		if alarms:
 			# query alarms. The "DESC Limit 3" means the query will return 3 frameID corresponding the highest summed score of individual seconds
-			query = "SELECT EventId, FrameId, TimeStamp, Type, SUM(Score) FROM Frames, Events WHERE Frames.Type='Alarm' AND Events.Id=Frames.EventId AND Frames.TimeStamp >= '" + st + "' AND Frames.Timestamp <= '" + et + "' AND MonitorId=" + m_id + " GROUP BY Timestamp ORDER BY sum(Score) DESC Limit 3;"
+			query = "SELECT EventId, FrameId, TimeStamp, Type, SUM(Score) FROM Frames, Events WHERE Frames.Type='Alarm' AND Events.Id=Frames.EventId AND Frames.TimeStamp >= '" + st + "' AND Frames.Timestamp <= '" + et + "' AND MonitorId=" + m_id + " GROUP BY Timestamp ORDER BY sum(Score) DESC Limit " + str(frame_analysis) + ";"
 			mycursor = mydb.cursor(MySQLdb.cursors.DictCursor)
 			mycursor.execute(query)
 			mydb.commit() # TRYING THIS HERE AS PER GOOGLE. May not be required.
@@ -260,6 +281,7 @@ def sql_query(retry):
 				# if it stills doesnt work, drop stale query
 				elif(retry >= 10):
 					printLog(tstamp(), "Stale query, drop alarm ", alarms[0])
+					print(myresult);
 					print(repr(query))
 					alarms.pop(0)
 					retry = 0
@@ -302,7 +324,6 @@ def ai_opencv4():
 	while(1):
 		# obj2 must be list in the form below for processing
 		#['car', 0.72265625, '/home/pic/13-561155-39.jpg', 'car', 0.73828125, '/home/pic/13-561155-46.jpg', 'car', 0.73828125, '/home/pic/13-561155-53.jpg']
-		
 		global use_email
 		obj = []
 		obj1 = []
@@ -403,21 +424,39 @@ def ai_opencv4():
 				if os.path.exists("/var/www/html/no_email"):
 					printLog(tstamp(), "email override enabled: /var/www/html/no_email")
 				elif use_email:
-					# convert lists to string
-					subject = " ".join(subject)
-					attachment = " ".join(attachment)
-					# prepare mail command
-					mail_string = "echo \"" + subject + "\" | mail --mime -s \"" + subject + "\" " + attachment + " " + email
-					mailstat = os.system(mail_string)
-					if mailstat != 0:
-						printLog(tstamp(), "mail status failed, will retry")
-						deleteme = 0 # keep files and trying to send mail
+					# Verify if camera is set to send email using filename
+					match = re.search(r'(\d+)-', attachment[0])
+					if match:
+						cam_number = match.group(1)
+						# split the emails variable into list
+						email_numbers = emails.split(',') # e.g. ['3', '17', '20', '9', '8', '11']
+						# email if cam is on the email list
+						if str(cam_number) in email_numbers:
+							print(f"Number {cam_number} found in emails!")
+							# convert lists to string
+							subject = " ".join(subject)
+							attachment = " ".join(attachment)
+							# prepare mail command
+							mail_string = "echo \"" + subject + "\" | mail --mime -s \"" + subject + "\" " + attachment + " " + email
+							mailstat = os.system(mail_string)
+							if mailstat != 0:
+								printLog(tstamp(), "mail status failed, will retry")
+								deleteme = 0 # keep files and trying to send mail
 			# after detection, delete items from tmp
 			if deleteme:
 				for x in line_str:
-					os.remove(x)
-				os.remove(list_of_files[0])
-		
+					try:
+						os.remove(x)
+					except OSError as e:
+						# Handle the case when the file doesn't exist
+						print(f"Error removing file {x}: {e}")
+
+				if list_of_files:
+					try:
+						os.remove(list_of_files[0])
+					except OSError as e:
+						# Handle the case when the file doesn't exist
+						print(f"Error removing file {list_of_files[0]}: {e}")	
 		time.sleep(delay)
 		# return is not used
 		# return subject
@@ -428,14 +467,37 @@ def mailme(subject, attachment):
 	if subject:
 		printLog(tstamp(), " mail status:", os.system(mail_string))
 
+# Functio to print to console and log. Maintain log files to 100 lines or less.
 def printLog(*args, **kwargs):
 	global log_enable
-#	print(*args, **kwargs)
+	max_lines = 500  # Maximum number of lines in the log file
 	print("\b\033[32m*", "\033[0m", *args, **kwargs)
 	if log_enable:
-		with open(log_file,'a+') as file:
+		with open(log_file, 'r') as file:
+		# Read all lines from the file
+			lines = file.readlines()
+		# Keep only the last 100 lines
+		lines = lines[-max_lines:]
+		# Open the file in write mode to overwrite its content
+		with open(log_file, 'w') as file:
+			# Write the last 100 lines back to the file
+			file.writelines(lines)	
+			# Write the new log entry
 			print(*args, **kwargs, file=file)
-
+def heartbeat():
+	# update timestamp to tmp file. Use external shell script to verify it runs normally.
+	if hb_enable:
+		try:
+		# Open the file in 'a' (append) mode and immediately close it
+			with open(path_hb + "alarm_hb", 'a'):
+				os.utime(path_hb + "alarm_hb", None)
+		except Exception as e:
+			printLog(tstamp(), "Error touching heartbeat file: " + str(e))
+	# Verify if opencv4 function is running in multiprocessing. If not exit.
+	if not ai_opencv4.is_alive():
+		printLog(tstamp(), "Function ai_opencv4 failed. Exiting.")
+		sys.exit()
+		
 if __name__ == '__main__':
 	#
 	# Create a multiprocessing Process with function ai_opencv4() and start it
@@ -446,4 +508,3 @@ if __name__ == '__main__':
 	main()
 	# Wait for both subprocesses to finish before exiting. Not needed.
 	# ai_opencv4.join()
-
